requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Tinkoff Invest API (T-Investments SDK)
# Install with: pip install t-tech-investments --index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
t-tech-investments

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0

# Telegram Bot
python-telegram-bot>=20.0

# Data processing
sqlalchemy>=2.0.0
pandas-ta>=0.3.14b0

# Utilities
aiohttp>=3.9.0
